+++ Procedural City/ProceduralTile.cpp
@@ -27,8 +27,8 @@ irr::scene::ISceneNode* Procedural::Tile::generateHouses(irr::scene::ISceneManag
 	//Place a group of houses on a single tile
 	for(int z = -11; z <= 11; z += 11)
 	{
-		tb.addHouse(core::vector3df(11,0.15f,z), -90);
-		tb.addHouse(core::vector3df(-11,0.15f,z), 90);
+		tb.addHouse(core::vector3df(11.0f,0.15f,(float)z), -90.0f);
+		tb.addHouse(core::vector3df(-11.0f,0.15f,(float)z), 90.0f);
 	}
 
 	return tb.getNode();
@@ -87,8 +87,8 @@ scene::ISceneNode* Procedural::Tile::generateTwoLaneAndHouses(scene::ISceneManag
 	//Place a group of houses on each side of the two lane road
 	for(int z = -11; z <= 11; z += 11)
 	{
-		tb.addHouse(core::vector3df(11,0.15f,z), -90);
-		tb.addHouse(core::vector3df(-11,0.15f,z), 90);
+		tb.addHouse(core::vector3df(11.0f,0.15f,(float)z), -90.0f);
+		tb.addHouse(core::vector3df(-11.0f,0.15f,(float)z), 90.0f);
 	}
 
 	return tb.getNode();
@@ -101,7 +101,7 @@ scene::ISceneNode* Procedural::Tile::generateTwoToSixTAndHouses(scene::ISceneMan
 
 	for(int z = -11; z <= 11; z += 11)
 	{
-		tb.addHouse(core::vector3df(-11,0.15f,z), 90);
+		tb.addHouse(core::vector3df(-11.0f,0.15f,(float)z), 90.0f);
 	}
 
 	return tb.getNode();
